Bug 51047: When we're trying to add decimal point (float values) discount, the marketplace is not loading any values [Android, IOS, Web]

+++ cart/src/main/java/com/boost/cart/ui/details/DetailsFragment.kt
@@ -521,7 +521,7 @@ class DetailsFragment : BaseFragment(), DetailsFragmentListener {
     imagePreviewPopUpFragement.show((activity as CartActivity).supportFragmentManager, IMAGE_PREVIEW_POPUP_FRAGMENT)
   }
 
-  private fun getDiscountedPrice(price: Double, discountPercent: Int): Double {
+  private fun getDiscountedPrice(price: Double, discountPercent: Double): Double {
     return price - ((discountPercent / 100) * price)
   }
