+++ marketplace/src/main/java/com/boost/marketplace/ui/details/call_track/CallTrackingActivity.kt
@@ -465,7 +465,7 @@ class CallTrackingActivity :
 
     }
 
-    private fun getDiscountedPrice(price: Double, discountPercent: Int): Double {
+    private fun getDiscountedPrice(price: Double, discountPercent: Double): Double {
         return price - ((discountPercent / 100) * price)
     }
